backend/sub_apps/models_predictions/requirements.txt
python-multipart
python-dotenv
fastapi 
uvicorn
sqlalchemy 
psycopg2-binary
pydantic
pandas
numpy 
openpyxl
requests
scikit-learn
xgboost
ibm-cos-sdk
joblib